add short hostname to workaround java7
java7 buffer overflows when fqdn > 64 chars (which is the case on gce).
workaround for now by inserting short hostname into /etc/hosts localhost
entry. better potential fix is to move to java8. we know that works with
our vanilla apache charms, but will need more testing to make sure that's
good for bigtop.